Otus/hw_4/requirements.txt
pytest==3.10.1
selenium==4.30.0
allure-pytest==2.14.1
requests==2.32.3
